refactor: remove drivers entrypoints and move factory
