+++ packages/bentocache/benchmarks/onetier_get_key.ts
@@ -10,8 +10,8 @@ import { caching } from 'cache-manager'
 import { redisStore } from 'cache-manager-ioredis-yet'
 
 import { BentoCache } from '../index.js'
-import { redisDriver } from '../drivers/redis.js'
 import { bentostore } from '../src/bento_store.js'
+import { redisDriver } from '../src/drivers/redis.js'
 import { REDIS_CREDENTIALS } from '../test_helpers/index.js'
 
 const bench = new Bench()
