+++ packages/bentocache/src/drivers/knex.ts
@@ -1,6 +1,21 @@
 import type { Knex } from 'knex'
 
-import type { DatabaseAdapter, KnexConfig } from '../types/main.js'
+import { DatabaseDriver } from './database.js'
+import type { CreateDriverResult, DatabaseAdapter, KnexConfig } from '../types/main.js'
+
+/**
+ * Create a knex driver
+ * You will need to install the underlying database package (mysql2, pg, sqlite3, etc)
+ */
+export function knexDriver(options: KnexConfig): CreateDriverResult<DatabaseDriver> {
+  return {
+    options,
+    factory: (config: KnexConfig) => {
+      const adapter = new KnexAdapter(config)
+      return new DatabaseDriver(adapter, config)
+    },
+  }
+}
 
 /**
  * Knex adapter for the DatabaseDriver
