+++ packages/bentocache/src/drivers/redis.ts
@@ -1,7 +1,32 @@
 import { Redis as IoRedis } from 'ioredis'
+import type { RedisOptions as IoRedisOptions } from 'ioredis'
 
 import { BaseDriver } from './base_driver.js'
-import type { L2CacheDriver, RedisConfig } from '../types/main.js'
+import { RedisBus } from '../bus/drivers/redis_bus.js'
+import type {
+  BusOptions,
+  CreateBusDriverResult,
+  CreateDriverResult,
+  L2CacheDriver,
+  RedisConfig,
+} from '../types/main.js'
+
+/**
+ * Create a new cache redis driver
+ */
+export function redisDriver(options: RedisConfig): CreateDriverResult<Redis> {
+  return { options, factory: (config: RedisConfig) => new Redis(config) }
+}
+
+/**
+ * Create a new bus redis driver. It leverages the Pub/sub capabilities of Redis
+ * to sending messages between your different processes.
+ */
+export function redisBusDriver(
+  options: { connection: IoRedisOptions } & BusOptions,
+): CreateBusDriverResult {
+  return { options, factory: (config: IoRedisOptions) => new RedisBus(config) }
+}
 
 /**
  * Caching driver for Redis
